improve seeder and searchbar

+++ database/migrations/2024_08_19_075802_create_products_table.php
@@ -4,20 +4,18 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
-    public function up(): void
-    {
+    public function up(): void {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
             $table->text('description');
             $table->foreignId('subcategory_id')->constrained('sub_categories')->onDelete('cascade');
             $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
-            $table->decimal('price', 8, 2);
+            $table->integer('price'); // Cambiato da decimal a integer
             $table->foreignId('categorydetails_id')->nullable()->constrained('category_details')->onDelete('set null');
             $table->integer('stock_quantity');
             $table->timestamps();
@@ -27,8 +25,7 @@ return new class extends Migration
     /**
      * Reverse the migrations.
      */
-    public function down(): void
-    {
+    public function down(): void {
         Schema::dropIfExists('products');
     }
 };
